Move weights panel to shared directory
